+++ src/main/java/jscover/server/UriNotFound.java
@@ -1,0 +1,7 @@
+package jscover.server;
+
+public class UriNotFound extends RuntimeException {
+    public UriNotFound(String message, Throwable cause) {
+        super(message, cause);
+    }
+}