+++ History.md
@@ -1,7 +1,8 @@
 0.2.0-RC1 / 2012-??-??
 ==================
+  * Alter JSONDataMerger.mergeJSONCoverageStrings to accept and array of coverage strings to merge
   * Improve unloaded JavaScript parsing error logging issue https://github.com/tntim96/JSCover/issues/27
-  * Internal - Remove source from JSON
+  * Internal: Load original JS source from file-system instead of JSON for report viewing
 
 0.1.1 / 2012-12-02
 ==================